+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # redmine deployment - CHANGELOG
 
+## 2026-06-16 v1.1.0
+* **[add]** deployment detail page _(HTML show)_ summarizing the key deployment data
+* **[add]** related issues & related revisions on the detail page, shown as tabs
+* **[add]** deployment statistics page _(successful deployments per month & per author, by environment)_
+* **[add]** sidebar links _(overview & statistics)_ and a link from the deployments list to each detail page
+
 ## 2024-10-01 v1.0.0
 * **[add]** API endpoint
 * **[add]** branches-resolving for changesets